Separated the test into contextualized files. Removed some tests.
